refactor: update regex
