+++ targets/frontend/src/lib/regex.js
@@ -1,0 +1,2 @@
+export const passwordValidation =
+  /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*]).{12,32}$/;
