--- targets/frontend/src/pages/api/regex.js
@@ -1,2 +1,0 @@
-export const passwordValidation =
-  /(?=.*[a-z])(?=.*[A-Z])(?=.*d)(?=.*[$@$!#.])[A-Za-zd$@$!%*?&.]/;
